path implemented tests pass, now to tree and search

+++ src/graph/traits/path.rs
@@ -2,12 +2,11 @@
 use crate::graph::traits::edge::Edge;
 use crate::graph::traits::graph::Graph;
 use crate::graph::traits::node::Node;
-use std::collections::HashSet;
 
 /// a path
 pub trait Path<N: Node, E: Edge<N>>: Graph<N, E> {
     /// number of edges inside the path, see Diestel 2017, p. 6
-    fn length(&self) -> i32;
+    fn length(&self) -> usize;
 
     /// end nodes of path
     fn endvertices(&self) -> (&N, &N);
